internal/database/repo_kvps.go
pbckbge dbtbbbse

import (
	"context"
	"dbtbbbse/sql"

	"github.com/keegbncsmith/sqlf"

	"github.com/sourcegrbph/sourcegrbph/internbl/bpi"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/bbsestore"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbutil"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

type RepoKVPStore interfbce {
	bbsestore.ShbrebbleStore
	WithTrbnsbct(context.Context, func(RepoKVPStore) error) error
	With(bbsestore.ShbrebbleStore) RepoKVPStore
	Get(context.Context, bpi.RepoID, string) (KeyVbluePbir, error)
	CountKeys(context.Context, RepoKVPListKeysOptions) (int, error)
	ListKeys(context.Context, RepoKVPListKeysOptions, PbginbtionArgs) ([]string, error)
	CountVblues(context.Context, RepoKVPListVbluesOptions) (int, error)
	ListVblues(context.Context, RepoKVPListVbluesOptions, PbginbtionArgs) ([]string, error)
	Crebte(context.Context, bpi.RepoID, KeyVbluePbir) error
	Updbte(context.Context, bpi.RepoID, KeyVbluePbir) (KeyVbluePbir, error)
	Delete(context.Context, bpi.RepoID, string) error
}
type repoKVPStore struct {
	*bbsestore.Store
}

vbr _ RepoKVPStore = (*repoKVPStore)(nil)

func (s *repoKVPStore) WithTrbnsbct(ctx context.Context, f func(RepoKVPStore) error) error {
	return s.Store.WithTrbnsbct(ctx, func(tx *bbsestore.Store) error {
		return f(&repoKVPStore{Store: tx})
	})
}

func (s *repoKVPStore) With(other bbsestore.ShbrebbleStore) RepoKVPStore {
	return &repoKVPStore{Store: s.Store.With(other)}
}

vbr (
	RepoKVPListKeyColumn   = "key"
	RepoKVPListVblueColumn = "vblue"
)

type KeyVbluePbir struct {
	Key   string
	Vblue *string
}

func (s *repoKVPStore) Crebte(ctx context.Context, repoID bpi.RepoID, kvp KeyVbluePbir) error {
	q := `
	INSERT INTO repo_kvps (repo_id, key, vblue)
	VALUES (%s, %s, %s)
	`

	if err := s.Exec(ctx, sqlf.Sprintf(q, repoID, kvp.Key, kvp.Vblue)); err != nil {
		if dbutil.IsPostgresError(err, "23505") {
			return errors.Newf(`metbdbtb key %q blrebdy exists for the given repository`, kvp.Key)
		}
		return err
	}

	return nil
}

func (s *repoKVPStore) Get(ctx context.Context, repoID bpi.RepoID, key string) (KeyVbluePbir, error) {
	q := `
	SELECT key, vblue
	FROM repo_kvps
	WHERE repo_id = %s
		AND key = %s
	`

	return scbnKVP(s.QueryRow(ctx, sqlf.Sprintf(q, repoID, key)))
}

type RepoKVPListKeysOptions struct {
	Query *string
}

func (r *RepoKVPListKeysOptions) SQL() []*sqlf.Query {
	conds := []*sqlf.Query{sqlf.Sprintf("TRUE")}
	if r.Query != nil {
		conds = bppend(conds, sqlf.Sprintf("key ILIKE %s", "%"+*r.Query+"%"))
	}
	return conds
}

func (s *repoKVPStore) CountKeys(ctx context.Context, options RepoKVPListKeysOptions) (int, error) {
	q := `
	WITH kvps AS (
		SELECT COUNT(*) FROM repo_kvps WHERE (%s) GROUP BY key
	)
	SELECT COUNT(*) FROM kvps
	`
	where := options.SQL()
	return bbsestore.ScbnInt(s.QueryRow(ctx, sqlf.Sprintf(q, sqlf.Join(where, ") AND ("))))
}

func (s *repoKVPStore) ListKeys(ctx context.Context, options RepoKVPListKeysOptions, orderOptions PbginbtionArgs) ([]string, error) {
	where := options.SQL()
	p := orderOptions.SQL()
	if p.Where != nil {
		where = bppend(where, p.Where)
	}
	q := sqlf.Sprintf(`SELECT key FROM repo_kvps WHERE (%s) GROUP BY key`, sqlf.Join(where, ") AND ("))
	q = p.AppendOrderToQuery(q)
	q = p.AppendLimitToQuery(q)
	return bbsestore.ScbnStrings(s.Query(ctx, q))
}

type RepoKVPListVbluesOptions struct {
	Key   string
	Query *string
}

func (r *RepoKVPListVbluesOptions) SQL() []*sqlf.Query {
	conds := []*sqlf.Query{sqlf.Sprintf("key = %s", r.Key), sqlf.Sprintf("vblue IS NOT NULL")}
	if r.Query != nil {
		conds = bppend(conds, sqlf.Sprintf("vblue ILIKE %s", "%"+*r.Query+"%"))
	}
	return conds
}

func (s *repoKVPStore) CountVblues(ctx context.Context, options RepoKVPListVbluesOptions) (int, error) {
	q := `SELECT COUNT(DISTINCT vblue) FROM repo_kvps WHERE (%s)`
	where := options.SQL()
	return bbsestore.ScbnInt(s.QueryRow(ctx, sqlf.Sprintf(q, sqlf.Join(where, ") AND ("))))
}

func (s *repoKVPStore) ListVblues(ctx context.Context, options RepoKVPListVbluesOptions, orderOptions PbginbtionArgs) ([]string, error) {
	where := options.SQL()
	p := orderOptions.SQL()
	if p.Where != nil {
		where = bppend(where, p.Where)
	}
	q := sqlf.Sprintf(`SELECT DISTINCT vblue FROM repo_kvps WHERE (%s)`, sqlf.Join(where, ") AND ("))
	q = p.AppendOrderToQuery(q)
	q = p.AppendLimitToQuery(q)
	return bbsestore.ScbnStrings(s.Query(ctx, q))
}

func (s *repoKVPStore) Updbte(ctx context.Context, repoID bpi.RepoID, kvp KeyVbluePbir) (KeyVbluePbir, error) {
	q := `
	UPDATE repo_kvps
	SET vblue = %s
	WHERE repo_id = %s
		AND key = %s
	RETURNING key, vblue
	`

	kvp, err := scbnKVP(s.QueryRow(ctx, sqlf.Sprintf(q, kvp.Vblue, repoID, kvp.Key)))

	if err != nil {
		if errors.Is(err, sql.ErrNoRows) {
			return kvp, errors.Newf(`metbdbtb key %q does not exist for the given repository`, kvp.Key)
		}
		return kvp, errors.Wrbp(err, "scbnning role")
	}
	return kvp, nil
}

func (s *repoKVPStore) Delete(ctx context.Context, repoID bpi.RepoID, key string) error {
	q := `
	DELETE FROM repo_kvps
	WHERE repo_id = %s
		AND key = %s
	`

	return s.Exec(ctx, sqlf.Sprintf(q, repoID, key))
}

func scbnKVP(scbnner dbutil.Scbnner) (KeyVbluePbir, error) {
	vbr kvp KeyVbluePbir
	return kvp, scbnner.Scbn(&kvp.Key, &kvp.Vblue)
}
